ai/study-assistant-lab/requirements.txt
openai>=1.9.0
python-dotenv
pydantic
pytest